+++ src/main/java/utils/ExtractorUtil.java
@@ -13,6 +13,8 @@ import java.io.FileNotFoundException;
 import java.io.IOException;
 import java.io.InputStream;
 import java.nio.charset.StandardCharsets;
+import java.time.DateTimeException;
+import java.util.Enumeration;
 import java.util.HashMap;
 import java.util.Map;
 
@@ -23,6 +25,8 @@ import org.apache.commons.compress.archivers.ArchiveStreamFactory;
 import org.apache.commons.compress.archivers.StreamingNotSupportedException;
 import org.apache.commons.compress.archivers.sevenz.SevenZArchiveEntry;
 import org.apache.commons.compress.archivers.sevenz.SevenZFile;
+import org.apache.commons.compress.archivers.zip.ZipArchiveEntry;
+import org.apache.commons.compress.archivers.zip.ZipFile;
 import org.apache.commons.compress.compressors.gzip.GzipCompressorInputStream;
 import org.codehaus.plexus.util.IOUtil;
 
@@ -34,9 +38,49 @@ public final class ExtractorUtil {
 
 	public static Map<String, String> unzip(File file) {
 		Logging.info("ExtractorUtil: starting extract");
-		Map<String, String> files = new HashMap<>();
+		Map<String, String> files = null;
 		String archiveFormat = detectArchiveFormat(file);
-		try (ArchiveInputStream ais = new ArchiveStreamFactory().createArchiveInputStream(archiveFormat,
+		if (ArchiveStreamFactory.ZIP.equals(archiveFormat)) {
+			files = extractZIP(file);
+		} else {
+			files = extractArchive(file, archiveFormat);
+		}
+		return files;
+	}
+
+	private static String detectArchiveFormat(File file) {
+		String archiveFormat = "";
+		try {
+			archiveFormat = ArchiveStreamFactory.detect(retrieveInputStream(file));
+		} catch (ArchiveException e) {
+			if (file.getName().contains(".tar")) {
+				archiveFormat = ArchiveStreamFactory.TAR;
+			} else {
+				Logging.error("Unable to detect archive format for file " + file.getAbsolutePath(), e);
+			}
+		}
+		return archiveFormat;
+	}
+
+	private static Map<String, String> extractZIP(File file) {
+		Map<String, String> files = new HashMap<>();
+		try (ZipFile zipFile = new ZipFile(file)) {
+			Enumeration<ZipArchiveEntry> zipEntries = zipFile.getEntries();
+			while (zipEntries.hasMoreElements()) {
+				ZipArchiveEntry entry = zipEntries.nextElement();
+				if (!entry.isDirectory()) {
+					files.put(entry.getName(), IOUtil.toString(zipFile.getInputStream(entry)));
+				}
+			}
+		} catch (IOException e) {
+			Logging.error("Unable to read ZIP file " + file.getAbsolutePath(), e);
+		}
+		return files;
+	}
+
+	private static Map<String, String> extractArchive(File file, String archiveFormat) {
+		Map<String, String> files = new HashMap<>();
+		try (ArchiveInputStream<ArchiveEntry> ais = new ArchiveStreamFactory().createArchiveInputStream(archiveFormat,
 				retrieveInputStream(file))) {
 			ArchiveEntry entry = null;
 			while ((entry = ais.getNextEntry()) != null) {
@@ -54,23 +98,14 @@ public final class ExtractorUtil {
 			Logging.error("Archive format is unknown " + archiveFormat, e);
 		} catch (IOException e) {
 			Logging.error("Unable to read zip file " + file.getAbsolutePath(), e);
+		} catch (DateTimeException e) {
+			// In version 1.25.0 (and earlier) of Apache Commons Compress lib
+			// the `DateTimeException` might occur indicating that PAX header
+			// of an archive is corrupt. We catch this manually until new lib's
+			// version comes out, which does not require us to do so.
+			Logging.error("Corrupted PAX header", e);
 		}
-
 		return files;
-	}
-
-	private static String detectArchiveFormat(File file) {
-		String archiveFormat = "";
-		try {
-			archiveFormat = ArchiveStreamFactory.detect(retrieveInputStream(file));
-		} catch (ArchiveException e) {
-			if (file.getName().contains(".tar")) {
-				archiveFormat = ArchiveStreamFactory.TAR;
-			} else {
-				Logging.error("Unable to detect archive format for file " + file.getAbsolutePath(), e);
-			}
-		}
-		return archiveFormat;
 	}
 
 	private static InputStream retrieveInputStream(File file) {